+++ tests/unit/PaymentUtilsTest.ts
@@ -1,7 +1,7 @@
 import type {OnyxEntry} from 'react-native-onyx';
 import type {BankAccountMenuItem} from '@components/Search/types';
 import {setPersonalBankAccountContinueKYCOnSuccess} from '@libs/actions/BankAccounts';
-import {approveMoneyRequest} from '@libs/actions/IOU';
+import {approveMoneyRequest} from '@libs/actions/IOU/ReportWorkflow';
 import Navigation from '@libs/Navigation/Navigation';
 import {getActivePaymentType, getBusinessBankAccountOptions, handleUnvalidatedAccount, selectPaymentType} from '@libs/PaymentUtils';
 import type {SelectPaymentTypeParams} from '@libs/PaymentUtils';
